requirements.txt
python-dotenv
aiogram
openai==0.28

